Redes_convolucionales/requirements.txt
tensorflow
torch
numpy
pandas
matplotlib
scikit-learn
Flask
Pillow
